+++ other/bootstrap_daemon/docker/tox-bootstrapd.sha256
@@ -1,1 +1,1 @@
-bded6f7ca320d8dfcb123a02c2c06aa9615b0e29e1d1d5b33b94bf88e85524d3  /usr/local/bin/tox-bootstrapd
+a12aa241a079e5f014a6689e48905a5a32c2fd455676cad431773908bda9245c  /usr/local/bin/tox-bootstrapd
